fix(udprep): don't trigger numba JIT just to import pure helpers (review #1)
Reaching udprep._radiation_compute (or RadiationSection) eagerly imported the
numba-JIT direct-shortwave kernels via udprep/__init__.py and a top-level import
in udprep_radiation, which stalls when __pycache__ isn't writable. Now:
- udprep/__init__.py exposes its names lazily (PEP 562) — `import udprep` is instant;
- udprep_radiation imports DirectShortwaveSolver lazily (only when the direct-SW
  path runs), keeping the kernels out of module import.
Making __init__ lazy exposed a latent udprep <-> section import cycle, so the
Section/SectionSpec base classes move to udprep/_section.py (sections import from
there; udprep.py re-exports for compatibility). `from udprep import
_radiation_compute` is now 0.00s and test_radiation_compute runs without
NUMBA_DISABLE_JIT. 278 tests, only the 2 pre-existing env failures.
