+++ tools/python/udprep/udprep_forcing.py
@@ -13,7 +13,7 @@ from typing import Any, Dict, List, Optional
 
 import numpy as np
 from scipy.interpolate import CubicSpline
-from .udprep import Section, SectionSpec
+from ._section import Section, SectionSpec
 
 DEFAULTS: Dict[str, Any] = Section.load_defaults_json().get("forcing", {})
 FIELDS: List[str] = list(DEFAULTS.keys())
